+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 ## Added
 - `SPNTA.marginalized_maxpost_params` property and `SPNTA.get_marginalized_param_sample()` method
+- Script to plot chains
 ## Changed
 ## Fixed
 - `SPNTA.get_marginalized_gp_noise_realization()` not correctly draws samples from the conditional distribution.