+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## Added
 - `SPNTA.marginalized_param_default_values` and `SPNTA.marginalized_maxpost_params` properties and `SPNTA.get_marginalized_param_sample()` method
 - Script to plot chains
+- In `SPNTA.save_new_parfile()`, also update the analytically marginalized parameter values and uncertainties.
+- `SPNTA.save_results()` also saves the analytically marginalized parameter values and uncertainties.
 ## Changed
 ## Fixed
 - `SPNTA.get_marginalized_gp_noise_realization()` now correctly draws samples from the conditional distribution.