+++ CHANGELOG.md
@@ -1,10 +1,21 @@
 # Unreleased
 ## Added
+- `SPNTA.marginalized_param_default_values` and `SPNTA.marginalized_maxpost_params` properties and `SPNTA.get_marginalized_param_sample()` method
+- Script to plot chains
+- In `SPNTA.save_new_parfile()`, also update the analytically marginalized parameter values and uncertainties.
+- `SPNTA.save_results()` also saves the analytically marginalized parameter values and uncertainties.
+- Maximum-posterior fitting in `VelaFitter`
+## Changed
+## Fixed
+- `SPNTA.get_marginalized_gp_noise_realization()` now correctly draws samples from the conditional distribution.
+## Removed
+
+# 0.1.0
+## Added
 - Can restart emcee from a prior chain
 - Script/function to resample existing chain
 - Saving autocorrelation lengths of chains
 - `SPNTA.has_ecorr` property
-- Maximum-posterior fitting in `VelaFitter`
 - Put the pulsar name as title in the result plot.
 - Default priors for T0, TASC, JUMP, M2, and H3.
 - Comments about the source of default priors.
@@ -15,6 +26,8 @@
 - `--analytic_marg` option in `pyvela` script
 - `--analytic_marg all` option in `pyvela` script
 - Add FD and FDJUMP to analytically marginalizable parameters
+- `SPNTA.maxpost_params` property
+- Pre-commit hooks
 ## Changed
 - Normalize the prior plot in `pyvela-plot` to improve visibility
 - Default PX prior now uses the diameter of the Galaxy as the maximum distance.
@@ -25,6 +38,9 @@
 - `pyvela` script now writes out a JLSO file into the output directory.
 - Read data from `JLSO` file while resuming `pyvela` script (also fixes parameter order).
 - Read data from `JLSO` file in `pyvela-rethin` script (also fixes handling of analytic marginalized parameters).
+- Updated documentation for analytic marginalization
+- In `pyvela` script, start one of the `emcee` walkers from the maximum-posterior point.
+- Updated CITATION file
 ## Fixed
 - Mark Julia lines wrongly missed in coverage with COV_EXCL_LINE to make the coverage % more accurate
 - Make `SPNTA.full_prior_dict()` work properly when "known" priors defined by Vela are used.